src/pyproject.toml
[project]
name = "authz-dashboard-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "pyyaml>=6.0.3",
]
